PlayerMoveLeft & Right Commands added
The commands for moving left and right. As per the previous version, there is a bit of refactoring that might be needed in terms of readability and maintainability, but I'm not sure how to do that. Awaiting code review of 2/11/2022 for that. Will be adding the PlayerMoveLeft.cs and PlayerMoveRight.cs soon.